Run instruction tests on all code generators
